requirements.txt
# 核心库
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0

# 数据获取
yfinance>=0.2.0
akshare>=1.0.0

# 深度学习
torch>=2.0.0

# 量化框架
pyqlib>=0.9.0

# NLP
transformers>=4.30.0

# Agent框架
langchain>=0.1.0
langchain-community>=0.1.0

# 大模型API
openai>=1.0.0

# 回测
backtrader>=1.9.0

# 前端
streamlit>=1.28.0
plotly>=5.17.0

# 工具
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.65.0
